Optimistic caching of tstat settings
When actuating the thermostat, Venstar requires that both setpoints be
set simultaneously. If you are setting one or the other but not both, you
want to use the cached value of what the other setpoint was last set to.
The problem is the cached value was only updated when the thermostat
reported, which is on some rate like ~10sec. If you send 2 commands in
close succession, the earlier one will be lost because the second
actuation command will result in the older cached value overwriting your
request.