Allow independent actuation of thermostat settings
Also prints out error messages in a few places where they weren't. Its a
bit unfortunate that we need to use float64 for deserializing the
messages, but we can can (and do) clamp fields to int where necessary.
This commit also switches bool -> int (0 and 1) for the override and fan
settings, which is a recent change to the xbos thrmsotat iface